+++ CHANGELOG.md
@@ -8,6 +8,7 @@ https://cdn.retail.red/omni/retailred-storefront-library-1.1.0.js
 - Introduced event `locationChanged` that is triggered whenever the customer changed the location.
 - Introduced `browserHistory` config to enable/disable the use of the browser history.
 - Fixed a bug that caused the modal to be removed after reaching the success page.
+- Introduced `customer.saveEmailAddress` config to remember the users email.
 
 ## v1.0.0
 https://cdn.retail.red/omni/retailred-storefront-library-1.0.0.js